// packages/oos-dsls-ts/src/generated/ast.ts
export interface AiHint extends langium.AstNode {
    readonly $container: DomainDecl;
    readonly $type: 'AiHint';
    body: string;
    name: string;
}

export const AiHint = {
    $type: 'AiHint',
    body: 'body',
    name: 'name'
} as const;

export function isAiHint(item: unknown): item is AiHint {
    return reflection.isInstance(item, AiHint.$type);
}

export interface AliasList extends langium.AstNode {
    readonly $container: DomainDecl;
    readonly $type: 'AliasList';
    values: Array<string>;
}

export const AliasList = {
    $type: 'AliasList',
    values: 'values'
} as const;
// ...
export interface DomainDecl extends langium.AstNode {
    readonly $container: DomainModel;
    readonly $type: 'DomainDecl';
    dsn: string;
    members: Array<DomainMember>;
    name: string;
    source: string;
}

export const DomainDecl = {
    $type: 'DomainDecl',
    dsn: 'dsn',
    members: 'members',
    name: 'name',
    source: 'source'
} as const;

export function isDomainDecl(item: unknown): item is DomainDecl {
    return reflection.isInstance(item, DomainDecl.$type);
}

export type DomainMember = AiHint | AliasList | Field | Meta | Permission | Relation;

export const DomainMember = {
    $type: 'DomainMember'
} as const;

export function isDomainMember(item: unknown): item is DomainMember {
    return reflection.isInstance(item, DomainMember.$type);
}

export interface DomainModel extends langium.AstNode {
    readonly $type: 'DomainModel';
    domain: DomainDecl;
}

export const DomainModel = {
    $type: 'DomainModel',
    domain: 'domain'
} as const;
// ...
export interface Example extends langium.AstNode {
    readonly $container: Field;
    readonly $type: 'Example';
    description: string;
    op: ExampleOp;
    value: ExampleValue;
}

export const Example = {
    $type: 'Example',
    description: 'description',
    op: 'op',
    value: 'value'
} as const;

export function isExample(item: unknown): item is Example {
    return reflection.isInstance(item, Example.$type);
}

export type ExampleOp = 'eq' | 'ge' | 'gt' | 'in' | 'le' | 'like' | 'lt' | 'ne';

export function isExampleOp(item: unknown): item is ExampleOp {
    return item === 'eq' || item === 'ne' || item === 'lt' || item === 'le' || item === 'gt' || item === 'ge' || item === 'like' || item === 'in';
}

export type ExampleValue = NumberValue | StringValue;

export const ExampleValue = {
    $type: 'ExampleValue'
} as const;
// ...
export interface Field extends langium.AstNode {
    readonly $container: DomainDecl;
    readonly $type: 'Field';
    examples: Array<Example>;
    modifiers: Array<FieldModifier>;
    name: string;
    type: FieldType;
}

export const Field = {
    $type: 'Field',
    examples: 'examples',
    modifiers: 'modifiers',
    name: 'name',
    type: 'type'
} as const;
// ...
export type FieldModifier = Filterable | OptionsRef | ReadOnly;

export const FieldModifier = {
    $type: 'FieldModifier'
} as const;
// ...
export type FieldType = 'boolean' | 'date' | 'datetime' | 'number' | 'string';

export function isFieldType(item: unknown): item is FieldType {
    return item === 'string' || item === 'number' || item === 'boolean' || item === 'date' || item === 'datetime';
}

export interface Filterable extends langium.AstNode {
    readonly $container: Field;
    readonly $type: 'Filterable';
}

export const Filterable = {
    $type: 'Filterable'
} as const;
// ...
export interface Meta extends langium.AstNode {
    readonly $container: DomainDecl;
    readonly $type: 'Meta';
    dsn?: string;
    labelField: string;
    name: string;
    orderBy?: string;
    table: string;
    valueField: string;
}

export const Meta = {
    $type: 'Meta',
    dsn: 'dsn',
    labelField: 'labelField',
    name: 'name',
    orderBy: 'orderBy',
    table: 'table',
    valueField: 'valueField'
} as const;

export function isMeta(item: unknown): item is Meta {
    return reflection.isInstance(item, Meta.$type);
}

export type NavModifier = As | Bind | Confirm;

export const NavModifier = {
    $type: 'NavModifier'
} as const;

export function isNavModifier(item: unknown): item is NavModifier {
    return reflection.isInstance(item, NavModifier.$type);
}

export interface NumberValue extends langium.AstNode {
    readonly $container: Example;
    readonly $type: 'NumberValue';
    value: number;
}

export const NumberValue = {
    $type: 'NumberValue',
    value: 'value'
} as const;

export function isNumberValue(item: unknown): item is NumberValue {
    return reflection.isInstance(item, NumberValue.$type);
}

export interface OptionsRef extends langium.AstNode {
    readonly $container: Field;
    readonly $type: 'OptionsRef';
    meta: langium.Reference<Meta>;
}

export const OptionsRef = {
    $type: 'OptionsRef',
    meta: 'meta'
} as const;
// ...
export interface Permission extends langium.AstNode {
    readonly $container: DomainDecl;
    readonly $type: 'Permission';
    actions: Array<PermissionAction>;
    role: string;
}

export const Permission = {
    $type: 'Permission',
    actions: 'actions',
    role: 'role'
} as const;

export function isPermission(item: unknown): item is Permission {
    return reflection.isInstance(item, Permission.$type);
}

export type PermissionAction = 'delete' | 'read' | 'write';
// ...
export interface ReadOnly extends langium.AstNode {
    readonly $container: Field;
    readonly $type: 'ReadOnly';
}

export const ReadOnly = {
    $type: 'ReadOnly'
} as const;

export function isReadOnly(item: unknown): item is ReadOnly {
    return reflection.isInstance(item, ReadOnly.$type);
}

export interface Relation extends langium.AstNode {
    readonly $container: DomainDecl;
    readonly $type: 'Relation';
    foreignField: string;
    kind: RelationKind;
    localField: string;
    name: string;
    target: string;
}

export const Relation = {
    $type: 'Relation',
    foreignField: 'foreignField',
    kind: 'kind',
    localField: 'localField',
    name: 'name',
    target: 'target'
} as const;

export function isRelation(item: unknown): item is Relation {
    return reflection.isInstance(item, Relation.$type);
}

export type RelationKind = 'belongs_to' | 'has_many' | 'has_one';
// ...
export interface StringValue extends langium.AstNode {
    readonly $container: Example;
    readonly $type: 'StringValue';
    value: string;
}

export const StringValue = {
    $type: 'StringValue',
    value: 'value'
} as const;